didn't like the way base dao's for each id type were coded with the obsolete names and such, so i broke them out into their own namespaces now... i'm not sure this is "better" yet tho
