+++ ProphetsWay.EFTools.Tests/ProphetsWay.EFTools.Tests.csproj
@@ -1,25 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net452;net461;net471;net472;net48;netcoreapp3.1</TargetFrameworks>
-
-    <IsPackable>false</IsPackable>
+	  <TargetFrameworks>
+		  net452;net46;net461;net471;net472;net48;
+		  netcoreapp2.1;netcoreapp3.1;
+		  net50
+	  </TargetFrameworks>
+	  
+	<IsPackable>false</IsPackable>
   </PropertyGroup>
 
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3" />
-    <PackageReference Include="coverlet.collector" Version="3.0.2">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
+	<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
+	<PackageReference Include="xunit" Version="2.4.1" />
+	<PackageReference Include="xunit.runner.visualstudio" Version="2.4.3" />
+	<PackageReference Include="coverlet.collector" Version="3.0.2">
+	  <PrivateAssets>all</PrivateAssets>
+	  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+	</PackageReference>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ProphetsWay.Example.DataAccess.EF\ProphetsWay.Example.DataAccess.EF.csproj" />
-    <ProjectReference Include="..\ProphetsWay.Example\ProphetsWay.Example.Tests\ProphetsWay.Example.Tests.csproj" />
+	<ProjectReference Include="..\ProphetsWay.Example.DataAccess.EF\ProphetsWay.Example.DataAccess.EF.csproj" />
+	<ProjectReference Include="..\ProphetsWay.Example\ProphetsWay.Example.Tests\ProphetsWay.Example.Tests.csproj" />
   </ItemGroup>
 
 </Project>
